# Tweak ImmutableDictionary.TryGetValue to improve throughput