A few tweaks to ImmutableDictionary to reduce the costs of lookups:
- Changed SortedInt32KeyNode to no longer be recursive, and then inlined it into its two call sites to avoid additional indirection and duplicative branches
- Avoid wrapping and unwrapping key value pairs when comparing the first item in a HashBucket

cc: @AArnott, @agocke 

Benchmark on my machine...

| Name              | Before (us) | After (us) | Improvement | 
|-------------------|-------------|------------|-------------| 
|        Lookup1Int | 175         | 117.7      | 32.74%      | 
|      Lookup100Int | 214.3       | 172.7      | 19.41%      | 
|    Lookup10000Int | 525.6       | 507.7      | 3.41%       | 
|     Lookup1String | 563.6       | 441.9      | 21.59%      | 
|   Lookup100String | 632.3       | 530.7      | 16.07%      | 
| Lookup10000String | 1,477.10  | 1,318.00 | 10.77%      | 

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Attributes.Jobs;
using BenchmarkDotNet.Running;
using System;
using System.Collections.Immutable;
using System.Runtime.CompilerServices;

[MemoryDiagnoser]
[InProcess]
public class Benchmark
{
    private static void Main() => BenchmarkRunner.Run<Benchmark>();

    private ImmutableDictionary<int, int> _oneInt;
    private ImmutableDictionary<int, int> _hundredInts;
    private ImmutableDictionary<int, int> _tenThousandInts;

    private ImmutableDictionary<string, string> _oneString;
    private ImmutableDictionary<string, string> _hundredStrings;
    private ImmutableDictionary<string, string> _tenThousandStrings;

    private string[] _strings;

    [GlobalSetup]
    public void Setup()
    {
        _strings = new string[10_000];
        for (int i = 0; i < _strings.Length; i++) _strings[i] = i.ToString();

        _oneInt = ImmutableDictionary.Create<int, int>().Add(0, 0);
        _oneString = ImmutableDictionary.Create<string, string>().Add(_strings[0], _strings[0]);

        _hundredInts = ImmutableDictionary.Create<int, int>();
        _hundredStrings = ImmutableDictionary.Create<string, string>();
        for (int i = 0; i < 100; i++)
        {
            _hundredInts = _hundredInts.Add(i, i);
            _hundredStrings = _hundredStrings.Add(_strings[i], _strings[i]);
        }

        _tenThousandInts = ImmutableDictionary.Create<int, int>();
        _tenThousandStrings = ImmutableDictionary.Create<string, string>();
        for (int i = 0; i < 10000; i++)
        {
            _tenThousandInts = _tenThousandInts.Add(i, i);
            _tenThousandStrings = _tenThousandStrings.Add(_strings[i], _strings[i]);
        }
    }

    [Benchmark]
    public int Lookup1Int()
    {
        int count = 0;
        for (int i = 0; i < 10000; i++)
        {
            for (int j = 0; j < 1; j++)
            {
                if (_oneInt.TryGetValue(j, out _))
                {
                    count++;
                }
            }
        }
        return count;
    }

    [Benchmark]
    public int Lookup100Int()
    {
        int count = 0;
        {
            for (int i = 0; i < 100; i++)
            {
                for (int j = 0; j < 100; j++)
                {
                    if (_hundredInts.TryGetValue(j, out _))
                    {
                        count++;
                    }
                }
            }
        }
        return count;
    }

    [Benchmark]
    public int Lookup10000Int()
    {
        int count = 0;
        for (int i = 0; i < 1; i++)
        {
            for (int j = 0; j < 10000; j++)
            {
                if (_tenThousandInts.TryGetValue(j, out _))
                {
                    count++;
                }
            }
        }
        return count;
    }

    [Benchmark]
    public int Lookup1String()
    {
        string[] strings = _strings;

        int count = 0;
        {
            for (int i = 0; i < 10000; i++)
            {
                for (int j = 0; j < 1; j++)
                {
                    if (_oneString.TryGetValue(strings[j], out _))
                    {
                        count++;
                    }
                }
            }
        }
        return count;
    }

    [Benchmark]
    public int Lookup100String()
    {
        string[] strings = _strings;

        int count = 0;
        for (int i = 0; i < 100; i++)
        {
            for (int j = 0; j < 100; j++)
            {
                if (_hundredStrings.TryGetValue(strings[j], out _))
                {
                    count++;
                }
            }
        }
        return count;
    }

    [Benchmark]
    public int Lookup10000String()
    {
        string[] strings = _strings;

        int count = 0;
        for (int i = 0; i < 1; i++)
        {
            for (int j = 0; j < 10000; j++)
            {
                if (_tenThousandStrings.TryGetValue(strings[j], out _))
                {
                    count++;
                }
            }
        }
        return count;
    }
}
```