@ahsonkhan 2019-03-05
> public class Benchmark

Should we add this perf test to https://github.com/dotnet/performance?

cc @jorive, @adamsitnik