@ahsonkhan 2019-03-06
> Yes, benchmarks should be added to the performance repository.

Opened an issue: https://github.com/dotnet/performance/issues/371

Thanks for labeling it. Its a good issue for up-for-grabs.