@bp74 2019-03-24
Thanks for this optimization. Is there a plan when version 1.6 will be released as a stable package to NuGet?